chore(base-setup): implement base express project
This contains the root route of the application, the package.json which
has all the dependencies to get and API route running. Dummy tests for
the application have also been implemented.

It also contains and ESLint config file which specifies the rules for
linting the codebase.

YARN is used for dependency management as I've found it to me much more
stable than NPM, so the yarn.lock file has been created also.