feat(get): implement the get cache by key endpoint
This endpoint follows the specifications in the document. It retrieves
the cache for a given key. In the event the key doesn't exist, it
generates a random value for that key, updates the cache and returns
the newly generated value.

The tests for this endpoint have also been implemented to ensure that
the value for a given key is returned and also that a new value is
generated for an invalid key.

TechDebt: remove test data seed from test suite